print the literal code nodes-to-string is choking on (#185)
Really dumb edit -- just pass down the content we're parsing from the top and do a slice (so as not to modify the original) on it. I was finding it challenging to read the AST and figure out what/where was breaking. With this, I can at least to a search in my package for the string.